Fix tests for Concourse compatibility
By default, Python gets tests from its `PYTHONPATH`.
The line which was removed now in each test forced to get tests
from the source code directory for developing purposes.

The reason why these lines have been removed is because
we are working with Concourse CI tool and we need the tests
to execute from the installed directory.