[dependency] Update python-dateutil to 2.8.0
This code updates the dependency of the python-dateutil package to 2.8.0 in the
setup.py and requirements.txt. The test `test_invalid_timezone` has been removed
since it is not needed anymore. The change was needed due to the fact that the
version 2.8.0 allows to deal with negative timezone offsets, something that
wasn't handled before.