[datetime] Handle dates with offsets outside 24h
When analyzing https://mail.gnome.org/archives/gtk-list/`, certain
message dates are converted by `str_to_datetime` to a datetime obj
with no timezome and a large offset. For instance, the date
`Sun, 28 Feb 1999 19:17:37 -7700 (EST)` is converted to
`datetime(1999, 2, 28, 19, 17, 37, tzinfo=tzoffset(None, -277200))`,
which has an offset not included between -timedelta(hours=24) and
timedelta(hours=24) and produces a ValueError.

This code handles such a case by checking that the offset is valid
(using the method `astimezone`). If the offset is not valid, the
ValueError exception is handled, the date is converted to UTC
and the offset is removed.

Tests have been added accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>