[tests] Restrict the workflow for a release tag
This commit adds a rule to restrict the tests workflow
to run when a release tag is pushed to the repository.

Whenever the maintainer uses the `publish` command,
a release commit and a release tag are pushed. This would
trigger the tests workflow twice resulting in redundant
workflows in the logs. This change is needed to avoid this.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>